Remove add collection from this PR
